Inline the field size calculations rather than using a parametrically polymorphic size calculation function
This will allow us in the future to emit specialized code that uses knowledge about the type of the field, such as using ad hoc functions for built-in types rather than having them implement `Serialize`/`Deserialize`. That, in turn, will allow us to have different encodings for the same Rust type, such as switching between a fixed-width encoding and a varint encoding for `u64`, depending on the user-specified type.