README: Rolling out a new asymmetric field: readers first, then writers
Original text:
> Once this change has been rolled out (at least to clients), we can safely promote the field to required in a subsequent change.

This phrasing suggested that it's okay to only roll out the change to the [writing] clients; but that would break any readers who don't know about the new schema.

So I added 2 more sentences to make the order of schema updates a bit clearer. Does this match your understanding?